Gee-Arr-Eff
This is from a day or two ago. I want to figure out the GRF format for Minecraft LCE.

I think the compressed GRF demo file is from Xbox 360, so I think it's data is formatted using XMemCompress.

For these demo scripts, I was trying to find the compression of the file, and also see if any of the data passes as valid NBT. I didn't find a compression format for the compressed file, out of the regular zlib-based formats, and I didn't find any NBT data either. I didn't expect to find any NBT there, but I was still curious if anything in there used NBT by chance.

https://github.com/PhoenixARC/-PCK-Studio
https://github.com/PhoenixARC/-OMI-Filetype-Library
https://github.com/Minecraft-Manipulator/Documentation/blob/main/LCE-GRF.md

https://www.youtube.com/watch?v=Zd6zTuklghY