Projectifying
Made things into more of a proper project organization. Going to try implementing my own parser in TS, compared to porting the other implementation from C#. I think I can reverse engineer how things are working, then write a spec for the file format standalone from an implementation of it. That's been the most helpful to me in writing my own new parser for files, I want to see how the file is laid out, not necessarily how one goes about parsing it. Sometimes seeing language-specific implementations of it can get a bit hectic, because it involves other things nearby that code which don't particularly have to do with the specific implementation you're looking to implement. That's why I've tended to keep things like file IO and other things like that, separate from the actual file manipulations themselves. I want to have the implementation of the format itself to be a separate thing from how you use that. It should work no matter where you want to use it. The minimum IO of the implementation should be the file you give it, as well as the file it gives you when serializing it again. It doesn't have to do the local handling for you. It is a bonus if it does do that, but I don't think it should require that step in order to be able to parse your file.