Broken Parsing
Trying to get the table of names parsed, but it isn't staying to the format I thought it was supposed to adhere to. This is following the 'GameRuleFileReader' implementation from the OMI library.

Part of the other issue is that I'm not sure whether my decompressed version of the demo has the initial header still or not, whether it's starting with the body with the first bytes.

The song of the day is "Actual Brutal Facts", been in my head as of lately. Listened to all of The Harmony Codex again, man I do like it a lot. Especially "What Life Brings".

Going to go for a run now, wanted to make sure to hydrate enough before, that's been my hindrance with some of my runs. You can have enough to eat, but not having enough water is just as energy-taking.

https://github.com/PhoenixARC/-OMI-Filetype-Library/blob/main/OMI%20Filetypes%20Library/Workers/GameRuleWorker/GameRuleFileReader.cs